+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * **[Feat]**: added support to disable Gutenberg Editor
 * **[Feat]**: added task() function to run simple Tasks
+* **[Feat]**: added support for stringified controller@method name 
 * **[Refactor]**: removed public router() method from App class
 * **[Refactor]**: refactored app() function
 * **[Refactor]**: refactored QueryBuilders and related models